requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.10.0
keras-tuner>=1.1.0

# Machine Learning
imbalanced-learn>=0.8.0
xgboost>=1.5.0
lightgbm>=3.3.0

# LLM & RAG
transformers>=4.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=0.27.0
torch>=1.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# Web App
streamlit>=1.10.0

# Utilities
tqdm>=4.60.0
joblib>=1.1.0
python-dotenv>=0.20.0
pyyaml>=6.0

# Development
jupyterlab>=3.0.0
ipython>=8.0.0
black>=22.0.0
isort>=5.10.0
pytest>=7.0.0